Perform an X-11 like decomposition with any (non integer) periodicity.

Perform an X-11 like decomposition with any (non integer) periodicity.

x11plus(
  y,
  period,
  mul = TRUE,
  trend.horizon = 6,
  trend.degree = 2,
  trend.kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube", "Uniform",
    "Triangular", "Epanechnikov", "Trapezoidal"),
  trend.asymmetric = c("CutAndNormalize", "Direct", "MMSRE"),
  seas.s0 = c("S3X3", "S3X1", "S3X5", "S3X9", "S3X15"),
  seas.s1 = c("S3X5", "S3X3", "S3X1", "S3X9", "S3X15"),
  extreme.lsig = 1.5,
  extreme.usig = 2.5
)

Arguments

y

input time-series.

period

Period of the seasonal component, any positive real number.

mul

Boolean indicating if the decomposition mode is multiplicative (TRUE).

trend.horizon

bandwidth of trend filters.

trend.degree

polynomial order in local trend model.

trend.kernel

kernel weights in objective function.

trend.asymmetric

truncation type for symmetric filter.

seas.s0

Seasonal filter for B5, C5, D5.

seas.s1

seasonal filter for B10, C10, D10.

extreme.lsig

lower boundary used for outlier correction in irregular.

extreme.usig

upper boundary used for outlier correction in irregular.

Value

An object of the class 'JD3_X11PLUS', containg the decomposition and the parameters

Examples

q<-x11plus(rjd3toolkit::ABS$X0.2.09.10.M, 12)